* grub-core/loader/i386/pc/linux.c (grub_cmd_linux): Fix compilation 	for 64-bit platforms.

+++ grub-core/loader/i386/pc/linux.c
@@ -231,8 +231,9 @@ grub_cmd_linux (grub_command_t cmd __attribute__ ((unused)),
     }
 
   grub_dprintf ("linux", "[Linux-%s, setup=0x%x, size=0x%x]\n",
-		grub_linux_is_bzimage ? "bzImage" : "zImage", real_size,
-		grub_linux16_prot_size);
+		grub_linux_is_bzimage ? "bzImage" : "zImage",
+		(unsigned) real_size,
+		(unsigned) grub_linux16_prot_size);
 
   relocator = grub_relocator_new ();
   if (!relocator)
